refactor: rename service packages, update document annotations for errors and http models test: fix methods name in EventListenerImplIntegrationTest
