+++ src/test/java/io/github/auditlistener/integration/EventListenerImplIntegrationTest.java
@@ -15,9 +15,9 @@ import io.github.auditlistener.model.elastic.ErrorDocument;
 import io.github.auditlistener.model.elastic.HttpDocument;
 import io.github.auditlistener.model.elastic.MethodDocument;
 import io.github.auditlistener.repository.EventRepository;
-import io.github.auditlistener.service.ElasticSearchServiceImpl;
-import io.github.auditlistener.service.EventListenerImpl;
-import io.github.auditlistener.service.KafkaServiceImpl;
+import io.github.auditlistener.service.impl.ElasticSearchServiceImpl;
+import io.github.auditlistener.service.impl.EventListenerImpl;
+import io.github.auditlistener.service.impl.KafkaServiceImpl;
 import io.github.auditlistener.utils.EventListenerTestUtils;
 import org.apache.http.HttpHost;
 import org.apache.kafka.clients.producer.KafkaProducer;
@@ -149,8 +149,8 @@ public class EventListenerImplIntegrationTest {
                     assertNotNull(document);
                     assertEquals(correlationId, document.getCorrelationId());
                     assertEquals("START", document.getEventType());
-                    assertEquals("DEBUG", document.getLogLevel());
-                    assertEquals("TestService.testMethod", document.getMethodName());
+                    assertEquals("DEBUG", document.getLevel());
+                    assertEquals("TestService.testMethod", document.getMethod());
                     assertNotNull(document.getArgs());
                     assertEquals("success", document.getResult());
                 });
